+++ src/xtopcom/xmbus/xevent_block.h
@@ -41,7 +41,7 @@ public:
     bool new_block{false};
     virtual int32_t get_class_type() const override {return xstatistic::enum_statistic_event_block;}
 private:
-    virtual int32_t get_object_size_real() const override {
+    virtual size_t get_object_size_real() const override {
         return sizeof(*this) + get_size(get_result_data());
     }
 };
