+++ src/xtopcom/xmbus/xevent_timer.h
@@ -32,7 +32,7 @@ public:
 
     virtual int32_t get_class_type() const override {return xstatistic::enum_statistic_event_timer;}
 private:
-    virtual int32_t get_object_size_real() const override {
+    virtual size_t get_object_size_real() const override {
         return sizeof(*this) + get_size(get_result_data());
     }
 };
@@ -65,7 +65,7 @@ public:
 
     virtual int32_t get_class_type() const override {return xstatistic::enum_statistic_event_chain_timer;}
 private:
-    virtual int32_t get_object_size_real() const override {
+    virtual size_t get_object_size_real() const override {
         return sizeof(*this) + get_size(get_result_data());
     }
 };
